Data_Science_Projects/Customer Behavior Segmentation/requirements.txt
# Customer Behavior Segmentation - Requirements

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database Connectivity
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0

# Data Processing
scipy>=1.10.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development
pytest>=7.4.0